requirements.txt
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
tortoise-orm[asyncpg]>=0.19.0,<0.21.0
asyncpg>=0.25.0,<0.30.0
pydantic[email]>=2.0.0,<2.6.0
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.5,<1.0.0
pyyaml>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0
aiohttp>=3.8.0,<4.0.0